+++ Application/Services/Infrastructure/IEmployeePersistence.cs
@@ -9,6 +9,6 @@ namespace Application.Services.Infrastructure
         Task UpdateEmail(Employee employee);
         Task UpdatePhone(Employee employee);
         Task UpdateSalary(Employee employee);
-        Task DeleteEmployees(int[] employeeIds);
+        Task DeleteEmployees(long[] employeeIds);
     }
 }
