+++ Domain/Aggregates/Employee.cs
@@ -6,7 +6,7 @@ namespace Domain.Aggregates
 #warning add aggregateroot that contains id
     public class Employee
     {
-        public int Id { get; }
+        public long Id { get; }
         public string Name { get; }
         public string Surname { get; }
         public Gender Gender { get; }
@@ -17,7 +17,7 @@ namespace Domain.Aggregates
         public DateTime CreatedAt { get; }
         public DateTime UpdatedAt { get; private set; }
 
-        public Employee(int id, string name, string surname, Gender gender, bool isManager, decimal salary, string email, PhoneNumber phoneNumber, DateTime createdAt, DateTime updatedAt)
+        public Employee(long id, string name, string surname, Gender gender, bool isManager, decimal salary, string email, PhoneNumber phoneNumber, DateTime createdAt, DateTime updatedAt)
         {
             Id = id;
             Name = name;
@@ -36,7 +36,7 @@ namespace Domain.Aggregates
 
         }
 
-        public static Employee Create(int id, string name, string surname, Gender gender, decimal salary, string email, PhoneNumber phoneNumber)
+        public static Employee Create(long id, string name, string surname, Gender gender, decimal salary, string email, PhoneNumber phoneNumber)
         {
             var employee = new Employee(
                 id: id,
