+++ Infrastructure/Persistence/DatabaseContext/Migrations/20231125140442_Initial.cs
@@ -13,7 +13,7 @@ namespace Infrastructure.Persistence.DatabaseContext.Migrations
                 name: "Employees",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false),
+                    Id = table.Column<long>(type: "bigint", nullable: false),
                     Name = table.Column<string>(type: "nvarchar(25)", maxLength: 25, nullable: false),
                     Surname = table.Column<string>(type: "nvarchar(25)", maxLength: 25, nullable: false),
                     Gender = table.Column<int>(type: "int", nullable: false),
