+++ Infrastructure/Persistence/DatabaseContext/Migrations/EmployeeDbContextModelSnapshot.cs
@@ -24,8 +24,8 @@ namespace Infrastructure.Persistence.DatabaseContext.Migrations
 
             modelBuilder.Entity("Domain.Aggregates.Employee", b =>
                 {
-                    b.Property<int>("Id")
-                        .HasColumnType("int");
+                    b.Property<long>("Id")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
@@ -66,8 +66,8 @@ namespace Infrastructure.Persistence.DatabaseContext.Migrations
                 {
                     b.OwnsOne("Domain.Values.PhoneNumber", "PhoneNumber", b1 =>
                         {
-                            b1.Property<int>("EmployeeId")
-                                .HasColumnType("int");
+                            b1.Property<long>("EmployeeId")
+                                .HasColumnType("bigint");
 
                             b1.Property<string>("MobilePhone")
                                 .IsRequired()
